+++ pyinaturalist/__init__.py
@@ -3,3 +3,7 @@
 __author__ = 'Nicolas Noé'
 __email__ = 'nicolas@niconoe.eu'
 __version__ = '0.7.0'
+
+DEFAULT_USER_AGENT = 'Pyinaturalist {version}'.format(version=__version__)
+
+user_agent = DEFAULT_USER_AGENT